requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
httpx